requirements.txt
Django==1.7
django-grappelli==2.6.3
django-guardian==1.2.5
django-jinja==1.3.1
django-reversion==1.8.5
django-taggit==0.12.3
markdown2==2.3.0
python-social-auth==0.2.2
django-pipeline==1.4.7
django-filter==0.9.2
djangorestframework==3.1.0
